+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/persistance/repository/ChannelRepository.kt
@@ -75,9 +75,32 @@ public interface ChannelRepository {
     public suspend fun selectChannels(channelCIDs: List<String>, forceCache: Boolean = false): List<Channel>
 
     /**
+     * Select channel by full channel ID [Channel.cid]
+     *
+     * @param cid A [Channel.cid] as query specification.
+     *
+     * @return A channel found in repository.
+     */
+    public suspend fun selectChannelByCid(cid: String): Channel?
+
+    /**
+     * Select channels by full channel IDs [Channel.cid]
+     *
+     * @param cids A list of [Channel.cid] as query specification.
+     *
+     * @return A list of channels found in repository.
+     */
+    public suspend fun selectChannelsByCids(cids: List<String>): List<Channel>
+
+    /**
+     * Read which channel cids need sync.
+     */
+    public suspend fun selectChannelCidsBySyncNeeded(limit: Int = NO_LIMIT): List<String>
+
+    /**
      * Read which channels need sync.
      */
-    public suspend fun selectChannelsSyncNeeded(): List<Channel>
+    public suspend fun selectChannelsSyncNeeded(limit: Int = NO_LIMIT): List<Channel>
 
     /**
      * Sets the Channel.deleteAt for a channel.
@@ -131,4 +154,8 @@ public interface ChannelRepository {
 
     @VisibleForTesting
     public fun clearChannelCache()
+
+    private companion object {
+        private const val NO_LIMIT: Int = -1
+    }
 }